Purge wasteful and confusing longs from MT19937.
The upper bits were never used, and zeroed after every operation.  They also
made people believe the seed is better on 64 bit archs, instead of having
the upper halves thrown away.

+++ crawl-ref/source/rng.h
@@ -2,7 +2,7 @@
 #define RNG_H
 
 void seed_rng();
-void seed_rng(long seed);
+void seed_rng(uint32_t seed);
 void push_rng_state();
 void pop_rng_state();
